fix(prompt): keep the message when the code export cannot represent its media
Three things, finishing the frontend separation.

Fix a data-loss bug in the code export. The export's query asks only for text, tool
calls and tool results, so a media part arrives as an empty object.
`promptContentPartSchema` is a discriminated union on `__typename`, and content is
parsed as a whole array — so that empty object failed the parse and took the entire
message with it. The caller catches, logs "Cannot convert message" and filters it
out, so a prompt carrying one image quietly lost that message's text and tool calls
from the generated snippet as well. Confirmed by test before fixing. Dropping just
the unrepresentable part keeps the rest; media is still absent from the snippet,
which is the known limitation, but losing the message around it was not intended.

Move the media schemas and converters into their own files. `promptSchemas.ts` keeps
one line for `AnyPart`, and `promptUtils.ts` keeps none of the eight media
converters. Neither file has much upstream churn, so this is about the diff being
readable rather than about conflict.

Take the test helper out of upstream's `spanDownloadUtils.test.ts`. The fork's REST
paths widened the generic on `authApiFetch.GET`, which changed which overload
TypeScript picks for an inline mock literal — so upstream's test genuinely does not
compile without a change, verified by reverting it. What remains is an import and
three one-line swaps, with the helper in a fixture the fork owns: +4/-12 rather than
+19/-12.

The frontend now touches 276 added lines in files upstream owns, from 824 when the
audit started, with 2617 lines in files that cannot conflict.
